+++ internal/fixtures/doc.go
@@ -1,0 +1,4 @@
+/*
+This is for test fixtures only.
+*/
+package fixtures