+++ pkg/util/io.go
@@ -1,0 +1,15 @@
+package util
+
+import (
+	"io"
+)
+
+type nopWriteCloser struct {
+	io.Writer
+}
+
+func (nopWriteCloser) Close() error { return nil }
+
+func NopWriteCloser(w io.Writer) io.WriteCloser {
+	return nopWriteCloser{w}
+}